requirements.txt
Pillow==10.1.0
requests==3.2.0
customtkinter==5.2.1
